Moved example follow.js to a folder crawler that test to follow many links and retrieve information. Changed to standard js.
